Revert "refactor(subcategories): implement useSubcategories to CategoryPage and SubcategoriesGrid #103"
This reverts commit 74c1394dd76ecbeef73a472745575e218d0ee787.
